+++ packages/plugin-packages/mashroom-monitoring-metrics-collector/type-definitions/api.ts
@@ -18,35 +18,38 @@ export interface Summary {
     observe(value: number, labels?: MetricLabels): void;
 }
 
+export type AggregationHint = 'omit' | 'sum' | 'first' | 'min' | 'max' | 'average';
+
 /**
  * Mashroom Monitoring Metrics Collector Service
  *
  * It uses the metric types defined here: https://prometheus.io/docs/concepts/metric_types
  * and supports also *labels* which can be used to differentiate the characteristics of the thing that is being measured;
  * e.g. to group requests total by the HTTP error code.
+ *
+ * The AggregationHint can be used if you need to aggregate metrics, e.g. in a Node.js cluster.
  */
 export interface MashroomMonitoringMetricsCollectorService {
-
     /**
      * A counter is a cumulative metric that represents a single monotonically increasing counter
      * whose value can only increase.
-     * If though the returned Counter has a set() method, the new value must always be higher then the current.
+     * If though the returned Counter has a set() method, the new value must always be higher than the current.
      */
-    counter(name: string, help: string): Counter;
+    counter(name: string, help: string, aggregationHint?: AggregationHint): Counter;
     /**
      * A gauge is a metric that represents a single numerical value that can arbitrarily go up and down.
      */
-    gauge(name: string, help: string): Gauge;
+    gauge(name: string, help: string, aggregationHint?: AggregationHint): Gauge;
     /**
      * A histogram samples observations (usually things like request durations or response sizes)
      * and counts them in configurable buckets.It also provides a sum of all observed values.
      */
-    histogram(name: string, help: string, buckets?: number[]): Histogram;
+    histogram(name: string, help: string, buckets?: number[], aggregationHint?: AggregationHint): Histogram;
     /**
      * Similar to a histogram, a summary samples observations. While it also provides a total count of
      * observations and a sum of all observed values, it calculates configurable quantiles..
      */
-    summary(name: string, help: string, quantiles?: number[]): Summary;
+    summary(name: string, help: string, quantiles?: number[], aggregationHint?: AggregationHint): Summary;
     /**
      * Get the collected metrics
      */
@@ -64,6 +67,7 @@ type MetricDataBase = {
 
 export type CounterMetricData = MetricDataBase & {
     type: 'counter';
+    aggregationHint: AggregationHint;
     data: Array<{
         value: number;
         labels: MetricLabels;
@@ -72,6 +76,7 @@ export type CounterMetricData = MetricDataBase & {
 
 export type GaugeMetricData = MetricDataBase & {
     type: 'gauge';
+    aggregationHint: AggregationHint;
     data: Array<{
         value: number;
         labels: MetricLabels;
@@ -80,6 +85,7 @@ export type GaugeMetricData = MetricDataBase & {
 
 export type HistogramMetricData = MetricDataBase & {
     type: 'histogram';
+    aggregationHint: AggregationHint;
     data: Array<{
         count: number;
         sum: number;
@@ -93,6 +99,7 @@ export type HistogramMetricData = MetricDataBase & {
 
 export type SummaryMetricData = MetricDataBase & {
     type: 'summary';
+    aggregationHint: AggregationHint;
     data: Array<{
         count: number;
         sum: number;