+++ packages/plugin-packages/mashroom-monitoring-prometheus-exporter/src/mashroom-bootstrap-webapp.ts
@@ -1,16 +1,26 @@
 
-import setupDefaultMetrics from './setup_default_metrics';
+import setupDefaultMetrics from './setup-default-metrics';
 import expressAppFactory from './webapp';
 
+import {startSyncRegistry, stopSyncRegistry} from './registry';
+import {startPM2Connector, stopPM2Connector} from './pm2-cluster-connector';
 import type { MashroomWebAppPluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
 
 const bootstrap: MashroomWebAppPluginBootstrapFunction = async (pluginName, pluginConfig, contextHolder) => {
-    const loggerFactory = contextHolder.getPluginContext().loggerFactory;
+    const pluginContext = contextHolder.getPluginContext();
+    const loggerFactory = pluginContext.loggerFactory;
     const { enableGcStats } = pluginConfig;
 
     setupDefaultMetrics(enableGcStats, loggerFactory);
 
-    return expressAppFactory(loggerFactory);
+    startSyncRegistry(pluginContext);
+    startPM2Connector(pluginContext);
+    pluginContext.services.core.pluginService.onUnloadOnce(pluginName, () => {
+        stopSyncRegistry();
+        stopPM2Connector();
+    });
+
+    return expressAppFactory();
 };
 
 export default bootstrap;